requirements-docker.txt

tqdm
psutil
termcolor
pandas
numpy
Pillow